+++ interfaces/inner_kits/native_cpp/camera_source/src/callback/dcamera_source_callback_stub.cpp
@@ -26,9 +26,9 @@ namespace OHOS {
 namespace DistributedHardware {
 DCameraSourceCallbackStub::DCameraSourceCallbackStub()
 {
-    memberFuncMap_[IDCameraSourceCBInterfaceCode::NOTIFY_REG_RESULT] =
+    memberFuncMap_[static_cast<uint32_t>(IDCameraSourceCBInterfaceCode::NOTIFY_REG_RESULT)] =
         &DCameraSourceCallbackStub::NotifyRegResultInner;
-    memberFuncMap_[IDCameraSourceCBInterfaceCode::NOTIFY_UNREG_RESULT] =
+    memberFuncMap_[static_cast<uint32_t>(IDCameraSourceCBInterfaceCode::NOTIFY_UNREG_RESULT)] =
         &DCameraSourceCallbackStub::NotifyUnregResultInner;
 }
